requirements.txt
# Core
torch>=2.4.0
numpy>=1.26.0
pyyaml>=6.0

# Simulation
git+ssh://git@github.com/ElyasYassin/robosuite.git

# RL
stable-baselines3[extra]>=2.0.0
gymnasium>=0.29.0

# Logging
tensorboard>=2.14.0
